feat: refactor + enable automatic task summary
