+++ packages/ai-terminal-assistant/src/browser/ai-terminal-assistant-service.ts
@@ -25,6 +25,8 @@ import * as monaco from '@theia/monaco-editor-core/esm/vs/editor/editor.api';
 import { LocalFileLinkProvider } from '@theia/terminal/lib/browser/terminal-file-link-provider';
 import { Widget } from '@theia/core/lib/browser';
 import { TaskTerminalWidgetManager } from '@theia/task/lib/browser/task-terminal-widget-manager';
+import { TaskWatcher } from '@theia/task/lib/common';
+import { TaskService } from '@theia/task/lib/browser/task-service';
 
 export interface SummaryRequest {
     cwd: string;
@@ -36,9 +38,14 @@ export const SummaryService = Symbol('SummaryService');
 
 export interface SummaryService {
     readonly onAllTerminalsClosed: Event<void>;
-    readonly onBuildFinished: Event<void>;
     readonly onCurrentTerminalBufferChanged: Event<void>;
+    readonly onSummaryRequestStarted: Event<void>;
+    readonly onSummaryRequestFinished: Event<Summary | undefined>;
+
+    currentTerminal: TerminalWidget;
     terminalBuffer: string[];
+    currentSummary: Summary | undefined;
+
     sendSummaryRequest(request: SummaryRequest): Promise<Summary | undefined>;
     sendSummaryRequestForLastUsedTerminal(): Promise<Summary | undefined>;
     openErrorInEditor(error: ErrorDetail): Promise<void>;
@@ -70,11 +77,20 @@ export class SummaryServiceImpl implements SummaryService {
     @inject(TaskTerminalWidgetManager)
     protected readonly taskTerminalWidgetManager: TaskTerminalWidgetManager;
 
+    @inject(TaskService)
+    protected readonly taskService: TaskService;
+
+    @inject(TaskWatcher)
+    protected readonly taskWatcher: TaskWatcher;
+
     protected readonly onAllTerminalsClosedEmitter = new Emitter<void>();
     readonly onAllTerminalsClosed: Event<void> = this.onAllTerminalsClosedEmitter.event;
 
-    protected readonly onBuildFinishedEmitter = new Emitter<void>();
-    readonly onBuildFinished: Event<void> = this.onBuildFinishedEmitter.event;
+    protected readonly onSummaryRequestStartedEmitter = new Emitter<void>();
+    readonly onSummaryRequestStarted: Event<void> = this.onSummaryRequestStartedEmitter.event;
+
+    protected readonly onSummaryRequestFinishedEmitter = new Emitter<Summary | undefined>();
+    readonly onSummaryRequestFinished: Event<Summary | undefined> = this.onSummaryRequestFinishedEmitter.event;
 
     protected readonly onCurrentTerminalBufferChangedEmitter = new Emitter<void>();
     readonly onCurrentTerminalBufferChanged: Event<void> = this.onCurrentTerminalBufferChangedEmitter.event;
@@ -83,28 +99,24 @@ export class SummaryServiceImpl implements SummaryService {
     protected editorDecorations: monaco.editor.IEditorDecorationsCollection | undefined;
     protected editorDisposables: monaco.IDisposable[] = [];
 
-    protected currentTerminal: TerminalWidget;
+    protected _currentTerminal: TerminalWidget;
+    get currentTerminal(): TerminalWidget {
+        return this._currentTerminal;
+    }
     terminalBuffer: string[] = [];
     protected hiddenTerminalContainer: HTMLDivElement | undefined;
     protected isTerminalVisible: boolean = false;
 
+    protected _currentSummary: Summary | undefined;
+    get currentSummary(): Summary | undefined {
+        return this._currentSummary;
+    }
+
+    protected summaryDebounce: ReturnType<typeof setTimeout> | undefined;
+
     @postConstruct()
     protected initialize(): void {
-        // Create a hidden container for the terminal widget
-        // This allows the terminal to be attached to the DOM (required for xterm.js)
-        // while keeping it invisible to the user
-        // We use clip-path to hide the content while keeping the element "visible" for phosphor
-        this.hiddenTerminalContainer = document.createElement('div');
-        this.hiddenTerminalContainer.id = 'ai-terminal-hidden-container';
-        this.hiddenTerminalContainer.style.position = 'fixed';
-        this.hiddenTerminalContainer.style.width = '800px';
-        this.hiddenTerminalContainer.style.height = '600px';
-        this.hiddenTerminalContainer.style.overflow = 'hidden';
-        // Use clip-path to make content invisible but keep element technically visible
-        // This ensures phosphor's isVisible check passes
-        // this.hiddenTerminalContainer.style.clipPath = 'inset(100%)';
-        this.hiddenTerminalContainer.style.pointerEvents = 'none';
-        document.body.appendChild(this.hiddenTerminalContainer);
+        this.initializeHiddenTerminal();
         // this.terminalService.all.forEach(terminal => {
         //     this.activeTerminals.add(terminal);
         //     terminal.onDidDispose(() => {
@@ -127,18 +139,70 @@ export class SummaryServiceImpl implements SummaryService {
 
         this.debugSessionManager.onDidStartDebugSession(async () => {
             console.log('Debug session started.');
+
         });
 
         this.debugSessionManager.onDidDestroyDebugSession(async () => {
             console.log('Debug session destroyed.');
-            this.onBuildFinishedEmitter.fire();
+            this.onSummaryRequestStartedEmitter.fire();
+            this._currentSummary = await this.sendSummaryRequestForLastUsedTerminal();
+            this.onSummaryRequestFinishedEmitter.fire(this.currentSummary);
+            this.createNewTerminal().catch(err => {
+                console.error('Error recreating hidden terminal after task exit:', err);
+            });
         });
 
         this.createNewTerminal().catch(err => {
             console.error('Error creating initial hidden terminal:', err);
         });
 
+        // Listen for task exit events to recreate the hidden terminal
+        // this.taskWatcher.onTaskExit(async () => {
+        //     this.onSummaryRequestStartedEmitter.fire();
+        //     this._currentSummary = await this.sendSummaryRequestForLastUsedTerminal();
+        //     this.onSummaryRequestFinishedEmitter.fire(this.currentSummary);
+        //     this.createNewTerminal().catch(err => {
+        //         console.error('Error recreating hidden terminal after task exit:', err);
+        //     });
+        // });
+
+        this.taskWatcher.onTaskExit(async () => {
+            clearTimeout(this.summaryDebounce);
+            this.summaryDebounce = setTimeout(async () => {
+                const running = await this.taskService.getRunningTasks();
+                if (running.length === 0) {
+                    // request summary
+                    this.onSummaryRequestStartedEmitter.fire();
+                    this._currentSummary = await this.sendSummaryRequestForLastUsedTerminal();
+                    this.onSummaryRequestFinishedEmitter.fire(this.currentSummary);
+                    this.createNewTerminal().catch(err => {
+                        console.error('Error recreating hidden terminal after task exit:', err);
+                    });
+                }
+            }, 500);
+        });
     }
+
+    protected initializeHiddenTerminal(): void {
+        // Initialization logic for the hidden terminal can be placed here if needed
+        // Create a hidden container for the terminal widget
+        // This allows the terminal to be attached to the DOM (required for xterm.js)
+        // while keeping it invisible to the user
+        // We use clip-path to hide the content while keeping the element "visible" for phosphor
+        this.hiddenTerminalContainer = document.createElement('div');
+        this.hiddenTerminalContainer.id = 'ai-terminal-hidden-container';
+        this.hiddenTerminalContainer.style.position = 'fixed';
+        this.hiddenTerminalContainer.style.width = '800px';
+        this.hiddenTerminalContainer.style.height = '600px';
+        this.hiddenTerminalContainer.style.overflow = 'hidden';
+        // Use clip-path to make content invisible but keep element technically visible
+        // This ensures phosphor's isVisible check passes
+        this.hiddenTerminalContainer.style.clipPath = 'inset(100%)';
+        this.hiddenTerminalContainer.style.pointerEvents = 'none';
+        document.body.appendChild(this.hiddenTerminalContainer);
+    }
+
+
 
     toggleTerminalVisibility(): void {
         if (this.hiddenTerminalContainer) {
@@ -154,12 +218,6 @@ export class SummaryServiceImpl implements SummaryService {
         }
     }
 
-    async createNewTaskTerminal(): Promise<void> {
-        const terminal = await this.taskTerminalWidgetManager.newTaskTerminal({ created: 'ai-terminal-assistant' });
-        this.terminalService.open(terminal, { mode: 'activate' });
-        await terminal.start();
-    }
-
     async createNewTerminal(): Promise<void> {
         // Dispose previous terminal if exists
         if (this.currentTerminal && !this.currentTerminal.isDisposed) {
@@ -169,14 +227,14 @@ export class SummaryServiceImpl implements SummaryService {
 
         const task = true;
         if (task) {
-            this.currentTerminal = await this.terminalService.newTerminal({
+            this._currentTerminal = await this.terminalService.newTerminal({
                 title: 'Hidden TaskTerminal',
                 destroyTermOnClose: true,
                 hideFromUser: true,
                 kind: 'task',
             });
         } else {
-            this.currentTerminal = await this.terminalService.newTerminal({
+            this._currentTerminal = await this.terminalService.newTerminal({
                 title: 'Hidden Terminal',
                 destroyTermOnClose: true,
                 hideFromUser: true,
